# openAPI.yaml
openapi: 3.0.3
info:
  description: "ft_transcendenceのAPI設計"
  title: "transcendence API"
  version: "v3.0"
paths:
  /users:
    get:
      description: 全ユーザーの情報を返します
      responses:
        200:
          description: 成功時のレスポンス
          content:
            application/json:
              schema:
                type: object
                properties:
                  usersList:
                    type: array
                    items:
                      $ref: "#/components/schemas/userIdAndName"
        500:
          description: 失敗時のレスポンス
  /users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
    get:
      description: ユーザーIDの情報を返します
      responses:
        200:
          description: ユーザーが存在した際のレスポンス
          content:
            application/json:
              schema:
                type: object
                properties:
                  userName:
                    type: string
                    example: "LabPixel"
                  isActive:
                    description: ユーザーがアクティブかどうかの判別
                    type: boolean
                    example: True
                  statistics:
                    $ref: '#/components/schemas/statistics'
                  friends:
                    description: フレンド一覧
                    type: array
                    items:
                      $ref: '#/components/schemas/userIdAndName'
                  blockingUsers:
                    description: ブロックしているユーザー一覧
                    type: array
                    items:
                      $ref: '#/components/schemas/userIdAndName'
                      
        400:
          description: ユーザーが存在しない際のレスポンス
        500:
          description: SQLなどのサーバーエラー
...
components:
  schemas:
...
    userIdAndName:
      description: ユーザーのIDと名前のオブジェクト
      properties:
        userId:
          description: "ユーザーID"
          type: integer
          example: 1
        userName:
          type: string
          example: "LabPixel"
        userIcon:
          description: ユーザーアイコンが格納されているパス
          type: string
          example: /assets/icon/1.ico
...
    statistics:
      description: 勝率などを格納するオブジェクト
      type: object
      properties:
        wins:
          description: 勝利数
          type: integer
          example: 10
        loses:
          description: 敗北数
          type: integer
          example: 10